Rename AppVeyorAuthFilter to something more generic
Also removes the log message that printed out the secret, yikes & uses a new AppSetting key for secret
